+++ pyproject.toml
@@ -14,6 +14,7 @@ dependencies = [
     "cachetools>=5.3.0",
     "prometheus-client>=0.19.0",
     "opentelemetry-sdk>=1.22.0",
+    "typing_extensions>=4.0.0", # Added for update_forward_refs
 ]
 
 [project.scripts]
