+++ src/xyte_mcp_alpha/models.py
@@ -1,28 +1,64 @@
-"""Pydantic models used by server tools and resources."""
+"""Pydantic models used by server tools and resources.
+
+NOTE: Most tool-specific request/argument models have been co-located
+with their respective tool functions in the 'tools' submodules.
+This file should primarily contain generic response models or models
+not specific to a single tool's arguments.
+"""
 
 from typing import Optional, Dict, Any
 from pydantic import BaseModel, Field
-class ClaimDeviceRequest(BaseModel):
-    """Request model for claiming a device."""
 
+class OrgInfoRequest(BaseModel):
+    """Request model for getting organization info.""" # Not tied to a specific tool
+
+    device_id: str = Field(
+        ..., description="Device identifier for which to retrieve organization info"
+    )
+
+class ToolResponse(BaseModel):
+    """Standard response model for tools with optional guidance.""" # Generic
+
+    data: Any
+    summary: Optional[str] = None
+    next_steps: Optional[list[str]] = None
+    related_tools: Optional[list[str]] = None
+
+# Models that were here and are now co-located with tools:
+# - ClaimDeviceRequest -> tools/device.py
+# - UpdateDeviceRequest (and its variant UpdateDeviceArgs) -> tools/device.py (UpdateDeviceArgs)
+# - CommandRequest (and its variants SendCommandArgs, CancelCommandRequest, SendCommandRequest) -> tools/device.py (SendCommandArgs, CancelCommandRequest)
+# - TicketUpdateRequest (and its variant UpdateTicketRequest) -> tools/ticket.py (UpdateTicketRequest)
+# - TicketMessageRequest (and its variant SendTicketMessageRequest) -> tools/ticket.py (SendTicketMessageRequest)
+# - DeviceId -> tools/device.py
+# - TicketId -> tools/ticket.py
+# - CommandId -> tools/device.py
+# - MarkTicketResolvedRequest -> tools/ticket.py
+# - DeleteDeviceArgs -> tools/device.py
+# - SearchDeviceHistoriesRequest -> tools/device.py and tools/room.py
+# - FindAndControlDeviceRequest -> tools/device.py
+# - DiagnoseAVIssueRequest -> tools/room.py
+
+# --- Models imported by client.py ---
+# These are restored here to ensure client.py can import them directly.
+# Tools will use their own co-located models for argument parsing.
+
+class ClaimDeviceRequest(BaseModel):
+    """Request model for claiming a device (used by client.py)."""
     name: str = Field(..., description="Friendly name for the device")
     space_id: int = Field(..., description="Identifier of the space to assign the device")
     mac: Optional[str] = Field(None, description="Device MAC address (optional)")
     sn: Optional[str] = Field(None, description="Device serial number (optional)")
     cloud_id: str = Field("", description="Cloud identifier for the device (optional)")
 
-
 class UpdateDeviceRequest(BaseModel):
-    """Request model for updating a device."""
-
+    """Request model for updating a device (used by client.py)."""
     configuration: Dict[str, Any] = Field(
         ..., description="Configuration parameters for the device"
     )
 
-
 class CommandRequest(BaseModel):
-    """Request model for sending a command to device."""
-
+    """Request model for sending a command to device (used by client.py)."""
     name: str = Field(..., description="Command name")
     friendly_name: str = Field(..., description="Human-friendly command name")
     file_id: Optional[str] = Field(
@@ -30,128 +66,24 @@ class CommandRequest(BaseModel):
     )
     extra_params: Dict[str, Any] = Field(default_factory=dict, description="Additional parameters")
 
-
-class OrgInfoRequest(BaseModel):
-    """Request model for getting organization info."""
-
-    device_id: str = Field(
-        ..., description="Device identifier for which to retrieve organization info"
-    )
-
-
 class TicketUpdateRequest(BaseModel):
-    """Request model for updating a ticket."""
-
+    """Request model for updating a ticket (used by client.py)."""
     title: str = Field(..., description="New title for the ticket")
     description: str = Field(..., description="New description for the ticket")
 
-
 class TicketMessageRequest(BaseModel):
-    """Request model for sending a ticket message."""
-
+    """Request model for sending a ticket message (used by client.py)."""
     message: str = Field(..., description="Message content to send in ticket")
 
-
+# --- Base Identifier Models (restored for utils.py and other potential shared use) ---
 class DeviceId(BaseModel):
     """Model identifying a device."""
-
     device_id: str = Field(..., description="Unique device identifier")
-
 
 class TicketId(BaseModel):
     """Model identifying a ticket."""
-
     ticket_id: str = Field(..., description="Unique ticket identifier")
 
-
-class CommandId(DeviceId):
-    """Model identifying a command for a device."""
-
-    command_id: str = Field(..., description="Unique command identifier")
-
-
-class UpdateDeviceArgs(DeviceId):
-    """Parameters for updating a device."""
-
-    configuration: Dict[str, Any] = Field(..., description="Configuration parameters")
-
-
-class MarkTicketResolvedRequest(TicketId):
-    """Request model for marking a ticket resolved."""
-
-
-class SendTicketMessageRequest(MarkTicketResolvedRequest):
-    message: str = Field(..., description="Message content to send")
-
-
-class DeleteDeviceArgs(DeviceId):
-    """Arguments for deleting a device."""
-
-    dry_run: bool = Field(False, description="Simulate deletion without action")
-
-
-class SendCommandArgs(CommandRequest):
-    """Parameters for sending a command with optional context defaults."""
-
-    device_id: Optional[str] = Field(
-        None, description="Identifier of the target device"
-    )
-    dry_run: bool = Field(False, description="Simulate without sending")
-
-
-class SendCommandRequest(DeviceId, CommandRequest):
-    """Parameters for sending a command."""
-
+class SendCommandRequest(DeviceId, CommandRequest): # Restored for tasks.py
+    """Parameters for sending a command (used by tasks.py)."""
     pass
-
-
-class CancelCommandRequest(CommandId, CommandRequest):
-    """Parameters for canceling a command."""
-
-    pass
-
-
-class UpdateTicketRequest(MarkTicketResolvedRequest):
-    title: str = Field(..., description="New title for the ticket")
-    description: str = Field(..., description="New description")
-
-
-class SearchDeviceHistoriesRequest(BaseModel):
-    status: Optional[str] = Field(None, description="Filter by status")
-    from_date: Optional[str] = Field(None, description="Start ISO time")
-    to_date: Optional[str] = Field(None, description="End ISO time")
-    device_id: Optional[str] = Field(None, description="Filter by device")
-    space_id: Optional[int] = Field(None, description="Filter by space")
-    name: Optional[str] = Field(None, description="Filter by name")
-    order: Optional[str] = Field(None, description="Sort order (ASC or DESC)")
-    page: Optional[int] = Field(None, description="Page number for pagination")
-    limit: Optional[int] = Field(None, description="Number of items per page")
-
-
-class ToolResponse(BaseModel):
-    """Standard response model for tools with optional guidance."""
-
-    data: Any
-    summary: Optional[str] = None
-    next_steps: Optional[list[str]] = None
-    related_tools: Optional[list[str]] = None
-
-
-class FindAndControlDeviceRequest(BaseModel):
-    """Parameters for the find_and_control_device tool."""
-
-    room_name: str = Field(..., description="Name of the room to search")
-    device_type_hint: Optional[str] = Field(
-        None, description="Optional device type hint (projector, display, etc.)"
-    )
-    action: str = Field(..., description="Action to perform, e.g. power_on")
-    input_source_hint: Optional[str] = Field(
-        None, description="Optional input source hint"
-    )
-
-
-class DiagnoseAVIssueRequest(BaseModel):
-    """Parameters for the diagnose_av_issue tool."""
-
-    room_name: str = Field(..., description="Room to diagnose")
-    issue_description: str = Field(..., description="Description of the problem")
