+++ src/xyte_mcp_alpha/server.py
@@ -12,20 +12,25 @@ sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "..",
 # Import everything using absolute imports
 import xyte_mcp_alpha.plugin as plugin
 from xyte_mcp_alpha.config import get_settings, validate_settings
-from xyte_mcp_alpha.events import GetNextEventRequest
+# from xyte_mcp_alpha.events import GetNextEventRequest # No longer directly used here
 from xyte_mcp_alpha.logging_utils import configure_logging, instrument
-import xyte_mcp_alpha.resources as resources
-import xyte_mcp_alpha.tools as tools
-import xyte_mcp_alpha.tasks as tasks
-import xyte_mcp_alpha.events as events
-import xyte_mcp_alpha.prompts as prompts
+# import xyte_mcp_alpha.resources as resources # Registrations moved
+# import xyte_mcp_alpha.tools as tools # Registrations moved
+import xyte_mcp_alpha.tasks as tasks # Still needed for /task/{task_id} route
+import xyte_mcp_alpha.events as events # Still needed for webhook, event stream, and GetNextEventRequest in events.py
+# import xyte_mcp_alpha.prompts as prompts # Registrations moved
+
+# Import registration functions
+from .resources import register_resources
+from .tools import register_tools
+from .prompts import register_prompts
 
 from prometheus_client import CONTENT_TYPE_LATEST, generate_latest
 from starlette.requests import Request
 from starlette.responses import Response, JSONResponse
 
 from mcp.server.fastmcp import FastMCP
-from mcp.types import ToolAnnotations
+# from mcp.types import ToolAnnotations # Registrations that used this are moved
 
 # Configure structured logging
 configure_logging()
@@ -122,151 +127,24 @@ async def list_resources_route(_: Request) -> JSONResponse:
     return JSONResponse(resources_list)
 
 
-# Resource registrations
-mcp.resource("devices://", description="List all devices")(
-    instrument("resource", "list_devices")(resources.list_devices)
-)
-mcp.resource(
-    "device://{device_id}/commands",
-    description="Commands issued to a device",
-)(instrument("resource", "list_device_commands")(resources.list_device_commands))
-mcp.resource(
-    "device://{device_id}/histories",
-    description="History records for a device",
-)(instrument("resource", "list_device_histories")(resources.list_device_histories))
-mcp.resource(
-    "device://{device_id}/status",
-    description="Current status of a device",
-)(instrument("resource", "device_status")(resources.device_status))
-mcp.resource(
-    "device://{device_id}/logs",
-    description="Recent logs for a device",
-)(instrument("resource", "device_logs")(resources.device_logs))
-mcp.resource(
-    "organization://info/{device_id}",
-    description="Organization info for a device",
-)(instrument("resource", "organization_info")(resources.organization_info))
-mcp.resource("incidents://", description="Current incidents")(
-    instrument("resource", "list_incidents")(resources.list_incidents)
-)
-mcp.resource("tickets://", description="All support tickets")(
-    instrument("resource", "list_tickets")(resources.list_tickets)
-)
-mcp.resource("ticket://{ticket_id}", description="Single support ticket")(
-    instrument("resource", "get_ticket")(resources.get_ticket)
-)
-mcp.resource(
-    "user://{user_token}/preferences",
-    description="Preferences for a user",
-)(resources.get_user_preferences)
-mcp.resource(
-    "user://{user_token}/devices",
-    description="Devices filtered by user",
-)(resources.list_user_devices)
+# Resource registrations (MOVED to resources.py)
 
-# Tool registrations
-mcp.tool(
-    description="Register a new device",
-    annotations=ToolAnnotations(readOnlyHint=False, destructiveHint=True),
-)(instrument("tool", "claim_device")(tools.claim_device))
-mcp.tool(
-    description="Remove a device from the organization",
-    annotations=ToolAnnotations(readOnlyHint=False, destructiveHint=True),
-)(instrument("tool", "delete_device")(tools.delete_device))
-mcp.tool(
-    description="Update configuration for a device",
-    annotations=ToolAnnotations(readOnlyHint=False, destructiveHint=True),
-)(instrument("tool", "update_device")(tools.update_device))
-mcp.tool(
-    description="Send a command to a device",
-    annotations=ToolAnnotations(readOnlyHint=False, destructiveHint=True),
-)(instrument("tool", "send_command")(tools.send_command))
-mcp.tool(
-    description="Cancel a previously sent command",
-    annotations=ToolAnnotations(readOnlyHint=False, destructiveHint=False),
-)(instrument("tool", "cancel_command")(tools.cancel_command))
-mcp.tool(
-    description="Update ticket details",
-    annotations=ToolAnnotations(readOnlyHint=False, destructiveHint=True),
-)(instrument("tool", "update_ticket")(tools.update_ticket))
-mcp.tool(
-    description="Resolve a ticket",
-    annotations=ToolAnnotations(readOnlyHint=False, destructiveHint=True),
-)(instrument("tool", "mark_ticket_resolved")(tools.mark_ticket_resolved))
-mcp.tool(
-    description="Send a message to a ticket",
-    annotations=ToolAnnotations(readOnlyHint=False, destructiveHint=False),
-)(instrument("tool", "send_ticket_message")(tools.send_ticket_message))
-mcp.tool(
-    description="Search device history records",
-    annotations=ToolAnnotations(readOnlyHint=True),
-)(instrument("tool", "search_device_histories")(tools.search_device_histories))
-mcp.tool(
-    description="Retrieve usage analytics for a device",
-    annotations=ToolAnnotations(readOnlyHint=True),
-)(instrument("tool", "get_device_analytics_report")(tools.get_device_analytics_report))
-mcp.tool(
-    description="Set context defaults",
-    annotations=ToolAnnotations(readOnlyHint=False, destructiveHint=False),
-)(instrument("tool", "set_context")(tools.set_context))
-mcp.tool(
-    description="Find and control a device with natural language hints",
-    annotations=ToolAnnotations(readOnlyHint=False, destructiveHint=True),
-)(instrument("tool", "find_and_control_device")(tools.find_and_control_device))
-mcp.tool(
-    description="Diagnose an AV issue in a room",
-    annotations=ToolAnnotations(readOnlyHint=True),
-)(instrument("tool", "diagnose_av_issue")(tools.diagnose_av_issue))
-mcp.tool(
-    description="Start meeting room preset",
-    annotations=ToolAnnotations(readOnlyHint=False, destructiveHint=True),
-)(instrument("tool", "start_meeting_room_preset")(tools.start_meeting_room_preset))
-mcp.tool(
-    description="Shutdown meeting room",
-    annotations=ToolAnnotations(readOnlyHint=False, destructiveHint=True),
-)(instrument("tool", "shutdown_meeting_room")(tools.shutdown_meeting_room))
-mcp.tool(
-    description="Log automation attempt",
-    annotations=ToolAnnotations(readOnlyHint=False, destructiveHint=False),
-)(instrument("tool", "log_automation_attempt")(tools.log_automation_attempt))
-mcp.tool(
-    description="Echo a message back",
-    annotations=ToolAnnotations(readOnlyHint=False, destructiveHint=False),
-)(instrument("tool", "echo_command")(tools.echo_command))
-mcp.tool(
-    description="Send a command asynchronously",
-    annotations=ToolAnnotations(readOnlyHint=False, destructiveHint=True),
-)(tasks.send_command_async)
-mcp.tool(
-    description="Get status of an asynchronous task",
-    annotations=ToolAnnotations(readOnlyHint=True),
-)(tasks.get_task_status)
+# Tool registrations (MOVED to tools/__init__.py)
+# Note: The registration for get_next_event_wrapper was here and is now also moved.
 
 
 @mcp.custom_route("/task/{task_id}", methods=["GET"])
-async def task_status(task_id: str) -> JSONResponse:
+async def task_status(task_id: str) -> JSONResponse: # tasks.get_task_status is still used here
     """Expose async task status via HTTP."""
     result = await tasks.get_task_status(task_id)
     return JSONResponse(result)
 
 
-# Create a wrapper function with explicit type annotation
-async def get_next_event_wrapper(params: GetNextEventRequest) -> Dict[str, Any]:
-    """Wrapper for get_next_event with explicit type annotation."""
-    return await events.get_next_event(params)
+# Create a wrapper function with explicit type annotation (MOVED to events.py)
 
+# Register the wrapper function as a tool (MOVED to tools/__init__.py)
 
-# Register the wrapper function as a tool
-mcp.tool(
-    description="Retrieve the next queued event",
-    annotations=ToolAnnotations(readOnlyHint=True),
-)(instrument("tool", "get_next_event")(get_next_event_wrapper))
-
-# Prompt registrations
-mcp.prompt()(prompts.reboot_device_workflow)
-mcp.prompt()(prompts.check_projectors_health)
-mcp.prompt()(prompts.proactive_projector_maintenance_check)
-mcp.prompt()(prompts.troubleshoot_offline_device_workflow)
+# Prompt registrations (MOVED to prompts.py)
 
 
 def get_server() -> Any:
@@ -274,10 +152,16 @@ def get_server() -> Any:
     settings = get_settings()
     validate_settings(settings)
 
+    # Register resources, tools, and prompts
+    register_resources(mcp)
+    register_tools(mcp)
+    register_prompts(mcp)
+
     plugin.load_plugins()
     if get_settings().enable_experimental_apis:
         from .experimental import echo
-
+        # Re-import ToolAnnotations if needed for this experimental tool, or move this registration too.
+        from mcp.types import ToolAnnotations # Added back for this specific registration
         mcp.tool(
             description="Echo a message (experimental)",
             annotations=ToolAnnotations(readOnlyHint=True),
