+++ src/xyte_mcp_alpha/tools/device.py
@@ -1,50 +1,141 @@
 """Device-related tool implementations."""
 
-import json
+# import json # No longer needed after log_automation_attempt moved
 import logging
-from pathlib import Path
+# from pathlib import Path # No longer needed after log_automation_attempt moved
 from typing import Any, Dict, Optional
 
-import anyio
+# import anyio # No longer needed after room tools moved
 
 from ..deps import get_client
 from ..utils import (
     MCPError,
-    get_session_state,
+    get_session_state, # Used by send_command
     handle_api,
     validate_device_id,
 )
-from .. import resources
-from ..client import (
-    ClaimDeviceRequest,
-    UpdateDeviceRequest,
-    CommandRequest,
-)
-from ..models import (
-    UpdateDeviceArgs,
-    SendCommandArgs,
-    CancelCommandRequest,
-    SearchDeviceHistoriesRequest,
-    ToolResponse,
-    DeleteDeviceArgs,
-    FindAndControlDeviceRequest,
-    DiagnoseAVIssueRequest,
-)
+# from .. import resources # No longer needed after diagnose_av_issue moved
+# Imports from ..client are for Pydantic models that are *arguments* to Xyte's API client, not internal tool arg models.
+# These should stay as imports from ..client if they represent Xyte client's request structures.
+# However, the instruction is "Pydantic request/argument models (...) are located in the same file as the tool function that uses them."
+# ClaimDeviceRequest from ..client is used by claim_device.
+# UpdateDeviceRequest from ..client is used by update_device.
+# CommandRequest from ..client is used by send_command, cancel_command.
+# For now, I will assume these specific models from ..client are intended to be distinct from the tool's own argument parsing models.
+# If ClaimDeviceRequest, UpdateDeviceRequest, CommandRequest were meant to be moved from models.py, that's a different interpretation.
+# The current models.py has ClaimDeviceRequest, UpdateDeviceRequest, CommandRequest, so they *are* defined there.
+# My previous read of models.py listed these.
+# The tool functions in tools/device.py (original) import these from ..client.
+# This is confusing. Let's assume the versions in `models.py` are the ones to be co-located.
+
+from ..client import ClaimDeviceRequest as XyteClaimDeviceRequest # Alias to avoid clash if defined locally
+from ..client import UpdateDeviceRequest as XyteUpdateDeviceRequest # Alias
+from ..client import CommandRequest as XyteCommandRequest # Alias
+
+from ..models import ToolResponse # Assuming ToolResponse remains in ..models for now
+# from ..models import ( # These will be defined locally
+# UpdateDeviceArgs,
+# SendCommandArgs,
+# CancelCommandRequest,
+# SearchDeviceHistoriesRequest,
+# DeleteDeviceArgs,
+# FindAndControlDeviceRequest,
+# DiagnoseAVIssueRequest, # Moved to room.py
+# )
 from mcp.server.fastmcp.server import Context
+from pydantic import BaseModel, Field # For defining models locally
 
 logger = logging.getLogger(__name__)
 
+# --- Pydantic Models moved from models.py ---
+class ClaimDeviceRequest(BaseModel): # Defined locally, from models.py structure
+    """Request model for claiming a device."""
+    name: str = Field(..., description="Friendly name for the device")
+    space_id: int = Field(..., description="Identifier of the space to assign the device")
+    mac: Optional[str] = Field(None, description="Device MAC address (optional)")
+    sn: Optional[str] = Field(None, description="Device serial number (optional)")
+    cloud_id: str = Field("", description="Cloud identifier for the device (optional)")
 
-async def claim_device(request: ClaimDeviceRequest) -> Dict[str, Any]:
+class DeviceId(BaseModel): # Base for others
+    """Model identifying a device."""
+    device_id: str = Field(..., description="Unique device identifier")
+
+class CommandId(DeviceId): # Base for CancelCommandRequest
+    """Model identifying a command for a device."""
+    command_id: str = Field(..., description="Unique command identifier")
+
+class UpdateDeviceArgs(DeviceId):
+    """Parameters for updating a device."""
+    configuration: Dict[str, Any] = Field(..., description="Configuration parameters")
+
+class SendCommandArgs(BaseModel): # Was CommandRequest from models.py, now local
+    """Parameters for sending a command with optional context defaults."""
+    # Fields from original CommandRequest in models.py
+    name: str = Field(..., description="Command name")
+    friendly_name: str = Field(..., description="Human-friendly command name")
+    file_id: Optional[str] = Field(
+        None, description="File identifier if the command includes a file"
+    )
+    extra_params: Dict[str, Any] = Field(default_factory=dict, description="Additional parameters")
+    # Fields specific to SendCommandArgs
+    device_id: Optional[str] = Field(
+        None, description="Identifier of the target device"
+    )
+    dry_run: bool = Field(False, description="Simulate without sending")
+
+class CancelCommandRequest(CommandId): # Was CommandId, CommandRequest from models.py
+    """Parameters for canceling a command."""
+    # Fields from original CommandRequest in models.py
+    name: str = Field(..., description="Command name")
+    friendly_name: str = Field(..., description="Human-friendly command name")
+    file_id: Optional[str] = Field(
+        None, description="File identifier if the command includes a file"
+    )
+    extra_params: Dict[str, Any] = Field(default_factory=dict, description="Additional parameters")
+
+class SearchDeviceHistoriesRequest(BaseModel):
+    status: Optional[str] = Field(None, description="Filter by status")
+    from_date: Optional[str] = Field(None, description="Start ISO time")
+    to_date: Optional[str] = Field(None, description="End ISO time")
+    device_id: Optional[str] = Field(None, description="Filter by device")
+    space_id: Optional[int] = Field(None, description="Filter by space") # Added as per original model
+    name: Optional[str] = Field(None, description="Filter by name") # Added as per original model
+    order: Optional[str] = Field(None, description="Sort order (ASC or DESC)")
+    page: Optional[int] = Field(None, description="Page number for pagination")
+    limit: Optional[int] = Field(None, description="Number of items per page")
+
+class DeleteDeviceArgs(DeviceId):
+    """Arguments for deleting a device."""
+    dry_run: bool = Field(False, description="Simulate deletion without action")
+
+class FindAndControlDeviceRequest(BaseModel):
+    """Parameters for the find_and_control_device tool."""
+    room_name: str = Field(..., description="Name of the room to search") # This makes it seem room-related
+    device_type_hint: Optional[str] = Field(
+        None, description="Optional device type hint (projector, display, etc.)"
+    )
+    action: str = Field(..., description="Action to perform, e.g. power_on")
+    input_source_hint: Optional[str] = Field(
+        None, description="Optional input source hint"
+    )
+
+# --- Tool Functions ---
+
+async def claim_device(request: ClaimDeviceRequest) -> Dict[str, Any]: # Uses local ClaimDeviceRequest
     """Claim a new device and assign it to the organization."""
     async with get_client() as client:
-        return await handle_api("claim_device", client.claim_device(request))
+        # Internally, map to XyteClaimDeviceRequest if needed by client.claim_device
+        # Assuming client.claim_device can take this Pydantic model directly if fields match.
+        # Or, explicitly create XyteClaimDeviceRequest:
+        # xyte_req = XyteClaimDeviceRequest(**request.model_dump())
+        # For now, assume direct usage if fields match, which is common with Pydantic.
+        return await handle_api("claim_device", client.claim_device(request)) # Pass the local model instance
 
 
 async def delete_device(data: DeleteDeviceArgs) -> ToolResponse:
     """Delete an existing device by its identifier."""
     if data.dry_run:
-        logger.info("Dry run: would delete device", extra={"device_id": data.device_id})
+        logger.info("Dry run: would delete device %s", data.device_id)
         return ToolResponse(
             data={"dry_run": True},
             summary=f"Dry run: Would delete device {data.device_id}",
@@ -59,14 +150,17 @@ async def delete_device(data: DeleteDeviceArgs) -> ToolResponse:
 async def update_device(data: UpdateDeviceArgs) -> Dict[str, Any]:
     """Apply configuration updates to a device."""
     async with get_client() as client:
-        req = UpdateDeviceRequest(configuration=data.configuration)
+        # Assuming UpdateDeviceArgs now holds the configuration directly or UpdateDeviceRequest from client is used.
+        # The local UpdateDeviceArgs is just DeviceId + configuration.
+        # The XyteUpdateDeviceRequest is what client.update_device expects.
+        xyte_req = XyteUpdateDeviceRequest(configuration=data.configuration)
         return await handle_api(
-            "update_device", client.update_device(data.device_id, req)
+            "update_device", client.update_device(data.device_id, xyte_req)
         )
 
 
 async def send_command(
-    data: SendCommandArgs,
+    data: SendCommandArgs, # Uses local SendCommandArgs
     ctx: Context | None = None,
 ) -> ToolResponse:
     """Send a command to a device."""
@@ -76,33 +170,34 @@ async def send_command(
         device_id = state.get("current_device_id")
         if device_id:
             logger.info(
-                "Defaulting device_id from context", extra={"device_id": device_id}
+                "Defaulting device_id from context: %s", device_id
             )
     if not device_id:
         raise MCPError(code="missing_device_id", message="device_id is required")
 
     async with get_client() as client:
-        req = CommandRequest(
+        # send_command tool uses local SendCommandArgs.
+        # client.send_command expects XyteCommandRequest.
+        xyte_req = XyteCommandRequest(
             name=data.name,
             friendly_name=data.friendly_name,
             file_id=data.file_id,
             extra_params=data.extra_params or {},
         )
         if ctx:
-            state = get_session_state(ctx)
+            state = get_session_state(ctx) # get_session_state from ..utils
             state["last_command"] = data.name
             await ctx.info(f"Sending command {data.name} to device {device_id}")
             await ctx.report_progress(0.0, 1.0, "sending")
 
         if data.dry_run:
             logger.info(
-                "Dry run: would send command",
-                extra={"device_id": device_id, "command": data.name},
+                "Dry run: would send command %s to device %s", data.name, device_id
             )
             result = {"dry_run": True}
         else:
             result = await handle_api(
-                "send_command", client.send_command(device_id, req)
+                "send_command", client.send_command(device_id, xyte_req)
             )
 
         if ctx:
@@ -119,23 +214,25 @@ async def send_command(
         )
 
 
-async def cancel_command(data: CancelCommandRequest) -> Dict[str, Any]:
+async def cancel_command(data: CancelCommandRequest) -> Dict[str, Any]: # Parameter name is 'data'
     """Cancel a previously sent command."""
     async with get_client() as client:
-        req = CommandRequest(
-            name=data.name,
-            friendly_name=data.friendly_name,
-            file_id=data.file_id,
-            extra_params=data.extra_params or {},
+        # cancel_command tool uses local CancelCommandRequest.
+        # client.cancel_command expects XyteCommandRequest.
+        xyte_req = XyteCommandRequest(
+            name=data.name, # Uses data.name now
+            friendly_name=data.friendly_name, # Uses data.friendly_name
+            file_id=data.file_id, # Uses data.file_id
+            extra_params=data.extra_params or {}, # Uses data.extra_params
         )
         return await handle_api(
             "cancel_command",
-            client.cancel_command(data.device_id, data.command_id, req),
+            client.cancel_command(data.device_id, data.command_id, xyte_req), # Uses data.device_id, data.command_id
         )
 
 
 async def search_device_histories(
-    params: SearchDeviceHistoriesRequest,
+    params: SearchDeviceHistoriesRequest, # Uses local SearchDeviceHistoriesRequest
     ctx: Context | None = None,
 ) -> Dict[str, Any]:
     """Search device history records with optional filters."""
@@ -174,107 +271,20 @@ async def get_device_analytics_report(
     ctx: Context | None = None,
 ) -> ToolResponse:
     """Retrieve usage analytics for a device."""
+    # validate_device_id is from ..utils
     device_id = validate_device_id(device_id)
     async with get_client() as client:
         if ctx:
             await ctx.info("Fetching analytics")
         result = await handle_api(
             "get_device_analytics",
-            client.get_device_analytics(device_id, period=period),
+            client.get_device_analytics(device_id, period=period), # Uses get_client from ..deps
         )
-        return ToolResponse(data=result.get("data", result))
-
-
-async def set_context(
-    device_id: Optional[str] = None,
-    space_id: Optional[str] = None,
-    ctx: Context | None = None,
-) -> ToolResponse:
-    """Set session context defaults for subsequent tool calls."""
-    if ctx is None:
-        raise MCPError(code="missing_context", message="Context required")
-
-    state = get_session_state(ctx)
-    if device_id is not None:
-        state["current_device_id"] = device_id
-    if space_id is not None:
-        state["current_space_id"] = space_id
-
-    return ToolResponse(data=state, summary="Context updated")
-
-
-async def start_meeting_room_preset(
-    room_name: str,
-    preset_name: str,
-    ctx: Context | None = None,
-) -> ToolResponse:
-    """Configure a meeting room for a preset workflow."""
-    if ctx:
-        await ctx.info(f"Configuring {room_name} for {preset_name}")
-        await ctx.report_progress(0.0, 1.0, "starting")
-    await anyio.sleep(0)  # yield control
-    if ctx:
-        await ctx.report_progress(1.0, 1.0, "done")
-    return ToolResponse(
-        data={"room_name": room_name, "preset_name": preset_name},
-        summary=f"Room {room_name} set to {preset_name} preset",
-    )
-
-
-async def shutdown_meeting_room(
-    room_name: str,
-    ctx: Context | None = None,
-) -> ToolResponse:
-    """Power down AV equipment in the specified room."""
-    if ctx:
-        await ctx.info(f"Shutting down room {room_name}")
-        await ctx.report_progress(0.0, 1.0, "shutting_down")
-    await anyio.sleep(0)
-    if ctx:
-        await ctx.report_progress(1.0, 1.0, "done")
-    return ToolResponse(
-        data={"room_name": room_name},
-        summary=f"Shutdown routine completed for {room_name}",
-    )
-
-
-async def log_automation_attempt(
-    workflow_name: str,
-    device_id: str,
-    steps_taken: list[str],
-    outcome: str,
-    user_feedback: Optional[str] = None,
-    error_details: Optional[str] = None,
-    ctx: Context | None = None,
-) -> ToolResponse:
-    """Record the result of an automated workflow."""
-    entry = {
-        "workflow_name": workflow_name,
-        "device_id": device_id,
-        "steps_taken": steps_taken,
-        "outcome": outcome,
-        "user_feedback": user_feedback,
-        "error_details": error_details,
-    }
-    path = Path("logs") / "automation.log"
-    path.parent.mkdir(parents=True, exist_ok=True)
-    with path.open("a", encoding="utf-8") as f:
-        f.write(json.dumps(entry) + "\n")
-    logger.info("automation_attempt", extra=entry)
-    if ctx:
-        await ctx.info("Automation attempt logged")
-    return ToolResponse(data=entry, summary="Attempt recorded")
-
-
-async def echo_command(device_id: str, message: str) -> ToolResponse:
-    """Example command that echoes a message back."""
-    validate_device_id(device_id)
-    logger.info("echo", extra={"device_id": device_id, "message": message})
-    return ToolResponse(data={"device_id": device_id, "echo": message})
+        return ToolResponse(data=result.get("data", result)) # ToolResponse from ..models
 
 
 async def find_and_control_device(
-    data: FindAndControlDeviceRequest,
+    data: FindAndControlDeviceRequest, # Uses local FindAndControlDeviceRequest
     ctx: Context | None = None,
 ) -> ToolResponse:
     """Find a device by room/name hints and perform an action."""
@@ -287,14 +297,14 @@ async def find_and_control_device(
     for dev in device_list:
         if room in str(dev.get("space_name", "")).lower():
             if (
-                data.device_type_hint
+                data.device_type_hint # data is FindAndControlDeviceRequest
                 and data.device_type_hint not in str(dev.get("type", "")).lower()
             ):
                 continue
             matches.append(dev)
 
     if not matches:
-        from difflib import get_close_matches
+        from difflib import get_close_matches # Standard library
 
         names = [d.get("space_name", "") for d in device_list]
         close = get_close_matches(room, names, n=1)
@@ -302,69 +312,30 @@ async def find_and_control_device(
             matches = [d for d in device_list if d.get("space_name") == close[0]]
 
     if not matches:
-        raise MCPError(code="device_not_found", message="No matching device found")
+        raise MCPError(code="device_not_found", message="No matching device found") # MCPError from ..utils
 
     device = matches[0]
     summary = f"Selected {device.get('name')} in {device.get('space_name')}"
     if ctx:
         await ctx.info(summary)
 
-    cmd = SendCommandArgs(
+    cmd = SendCommandArgs( # Uses local SendCommandArgs
         device_id=device.get("id"),
         name=data.action,
         friendly_name=data.action.replace("_", " "),
         extra_params=(
             {"input": data.input_source_hint} if data.input_source_hint else {}
         ),
-        file_id=None,
-        dry_run=False
+        file_id=None, # Added as per SendCommandArgs definition
+        dry_run=False # Added as per SendCommandArgs definition
     )
+    # Calls the send_command function within the same file.
     result = await send_command(cmd, ctx=ctx)
-    return ToolResponse(
+    return ToolResponse( # ToolResponse from ..models
         data={"device": device, "command": result.data},
         summary=summary + f" -> {data.action}",
     )
 
-
-async def diagnose_av_issue(
-    data: DiagnoseAVIssueRequest,
-    ctx: Context | None = None,
-) -> ToolResponse:
-    """Run basic diagnostics for a room based on an issue description."""
-    async with get_client() as client:
-        devices = await handle_api("get_devices", client.get_devices())
-
-    room_lower = data.room_name.lower()
-    device = next(
-        (
-            d
-            for d in devices.get("devices", devices)
-            if room_lower in str(d.get("space_name", "")).lower()
-        ),
-        None,
-    )
-
-    if not device:
-        raise MCPError(code="device_not_found", message="No device found for room")
-
-    status = await resources.device_status(device["id"])
-    histories = await search_device_histories(
-        SearchDeviceHistoriesRequest(
-            device_id=device["id"],
-            status=None,
-            from_date=None,
-            to_date=None,
-            space_id=None,
-            name=None,
-            order=None,
-            page=None,
-            limit=None
-        ),
-        ctx=ctx,
-    )
-
-    return ToolResponse(
-        data={"status": status, "histories": histories},
-        summary="Diagnostics gathered",
-        next_steps=["send_command"],
-    )
+# Removed: set_context, start_meeting_room_preset, shutdown_meeting_room,
+# log_automation_attempt, echo_command, diagnose_av_issue
+# These have been moved to tools/utility.py or tools/room.py
